Increments version to 0.3.0
Updates the library version from 0.2.1 to 0.3.0.

This change also addresses a potential issue with hash table
re-initialization by incrementing an epoch counter and resetting the
hash table when the counter reaches a maximum value.

Additionally, the commit refactors the NUM compression attempt to skip if
it doesn't achieve sufficient compression, improving overall
performance. It also updates the documentation of inline functions
to clearly indicate input and output parameters.